refactor: return 두번 나오지 않도록 컴포넌트 빼내기
